Move definition of FullState and RequestState in global types file
