+++ _dev/src/components/catalog/summary/panel/full-sync-request.vue
@@ -33,7 +33,7 @@
 
 <script lang="ts">
 import {defineComponent} from 'vue';
-import {RequestState} from '@/store/modules/catalog/types';
+import {RequestState} from '@/store/types';
 
 export default defineComponent({
   name: 'FullSyncRequest',
